Project discontinued…
I have some concerns about potential attack vectors, that would allow
Facebook to steal private keys from the bookmarklet. In case the project
became popular and Facebook had an interest in investing time to do so,
this would render it ineffective. Too bad, but an interesting
thought- and code-experiment anyway.